checker/src/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
enochecker==0.4.2
enochecker-cli==0.7.0
enochecker-core==0.10.0
eventlet==0.30.2
Flask==1.1.2
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsons==1.4.2
MarkupSafe==1.1.1
pycryptodome==3.10.1
pymongo==3.11.3
requests==2.25.1
six==1.15.0
typish==1.9.2
urllib3==1.26.4
Werkzeug==1.0.1
pysctp
